Moved all functionality to ScPlugin
Created a ScPlugin class and moved the current getPlugins functionality there.
