+++ distribution/openrct2.d.ts
@@ -56,7 +56,10 @@ declare global {
      * Plugin writers should check if ui is available using `typeof ui !== 'undefined'`.
      */
     var ui: Ui;
-
+    /**
+     * APIs for managing the installed plugins
+     */
+    var pluginManager: PluginManager;
     /**
      * Registers the plugin. This may only be called once.
      * @param metadata Information about the plugin and the entry point.
@@ -172,12 +175,6 @@ declare global {
      * Core APIs for storage and subscriptions.
      */
     interface Context {
-
-        /**
-         * Gets all of the currently registered plugins
-         */
-        getPlugins(): PluginMetadata[];
-
         /**
          * Gets the current version of the plugin api. This is an integer that increments
          * by 1 every time a change to the plugin api is made.
@@ -4953,5 +4950,12 @@ declare global {
         getAllObjects(type: "banner"): BannerObject[];
         getAllObjects(type: "scenery_group"): SceneryGroupObject[];
         getAllObjects(type: "music"): LoadedObject[];
+    }
+
+    /**
+     * Interface to handle the plugin manager
+     */
+    interface PluginManager {
+        getPlugins(): PluginMetadata[];
     }
 }
